backend/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
PyYAML==6.0 